fix(src): rtnetlink: Treat IFA_F_DEPRECATED like deletion
When dynamic IPv6 addresses change, old addresses are not
deleted immediately but rather deprecated. This is seen
on rtnetlink as a RTM_NEWADDR with changed ifa_flags.
We never want to announce deprecated addresses, as they
are likely no longer valid, so treat a RTM_NEWADDR with
ifa_flags & IFA_F_DEPRECATED as if it was RTM_DELADDR.

Previously, all messages with ifa_flags & IFA_F_DEPRECATED
were silently skipped and as such, when an address changes
from valid to deprecated, it would remain in wsdd's data.